vue-automessage/api/requirements.txt
Flask==2.2.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
PyYAML==6.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
alembic==1.8.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
flask-marshmallow==0.14.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
passlib==1.7.4
pip==22.2.2
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-file==1.5.1
setuptools==57.0.0
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.11
wheel==0.36.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
